Mathcalc

+++ Lektion 2/src/Math.java
@@ -8,9 +8,10 @@ public class Math {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
         return a / b;
         
     }
-    
+
+
 }
